add file mixin.py in finance, replace mixin class
